int to long

+++ src/main/java/org/rev317/min/accessors/Client.java
@@ -34,7 +34,7 @@ public interface Client {
 
     int[] getMenuActionId();
 
-    int[] getMenuAction1();
+    long[] getMenuAction1();
 
     int[] getMenuAction2();
